+++ src/django/api/tests.py
@@ -5,7 +5,6 @@ import os
 
 from openpyxl import load_workbook
 
-from datetime import datetime
 from dateutil.relativedelta import relativedelta
 from unittest.mock import Mock, patch
 
@@ -147,7 +146,7 @@ class FacilityListCreateTest(APITestCase):
         fields = NonstandardField.objects.filter(
             contributor=self.user.contributor).values_list(
             'column_name', flat=True)
-        self.assertEquals(1, len(fields))
+        self.assertEqual(1, len(fields))
         self.assertIn('extra_1', fields)
 
     def test_creates_source(self):
@@ -828,7 +827,7 @@ class GeocodingTest(TestCase):
     @patch('api.geocoding.requests.get')
     def test_geocode_non_200_response(self, mock_get):
         mock_get.return_value = Mock(ok=True, status_code=400)
-        with self.assertRaisesRegexp(ValueError, '400'):
+        with self.assertRaisesRegex(ValueError, '400'):
             geocode_address('Noorbagh, Kaliakoir Gazipur Dhaka 1704', 'BD')
 
 
@@ -2977,7 +2976,7 @@ class DashboardListTests(APITestCase):
         # Ensure they are ordered newest first
         self.assertEqual(
             ['Third List', 'Second List', 'First List'],
-            [l['name'] for l in lists])
+            [d['name'] for d in lists])
 
     def test_superuser_can_list_own_lists(self):
         self.client.login(email=self.superuser_email,
@@ -3009,7 +3008,7 @@ class DashboardListTests(APITestCase):
         self.assertEqual(3, len(lists))
         self.assertEqual(
             ['Third List', 'Second List', 'First List'],
-            [l['name'] for l in lists])
+            [d['name'] for d in lists])
 
     def test_user_cannot_list_other_contributors_lists(self):
         # Regular users, even if they ask for lists by other
@@ -3029,16 +3028,17 @@ class DashboardListTests(APITestCase):
         self.assertEqual(3, len(lists))
         self.assertEqual(
             ['Third List', 'Second List', 'First List'],
-            [l['name'] for l in lists])
+            [d['name'] for d in lists])
 
     def test_user_can_view_own_lists(self):
         self.client.login(email=self.user_email,
                           password=self.user_password)
 
-        for l in [self.list, self.inactive_list, self.private_list]:
-            response = self.client.get('/api/facility-lists/{}/'.format(l.id))
+        for fac_list in [self.list, self.inactive_list, self.private_list]:
+            response = self.client.get(
+                '/api/facility-lists/{}/'.format(fac_list.id))
             self.assertEqual(200, response.status_code)
-            self.assertEqual(l.name, response.json()['name'])
+            self.assertEqual(fac_list.name, response.json()['name'])
 
     def test_user_cannot_view_others_lists(self):
         self.client.login(email=self.user_email,
@@ -3057,10 +3057,11 @@ class DashboardListTests(APITestCase):
         superuser_lists = [self.superlist]
         user_lists = [self.list, self.inactive_list, self.private_list]
 
-        for l in superuser_lists + user_lists:
-            response = self.client.get('/api/facility-lists/{}/'.format(l.id))
+        for fac_list in superuser_lists + user_lists:
+            response = self.client.get(
+                '/api/facility-lists/{}/'.format(fac_list.id))
             self.assertEqual(200, response.status_code)
-            self.assertEqual(l.name, response.json()['name'])
+            self.assertEqual(fac_list.name, response.json()['name'])
 
 
 class FacilityDeleteTest(APITestCase):
@@ -3782,7 +3783,7 @@ class FacilityMergeTest(APITestCase):
         self.facility_2_claim.status = FacilityClaim.APPROVED
         self.facility_2_claim.save()
 
-        just_before_change = datetime.utcnow()
+        just_before_change = timezone.now()
         self.client.login(email=self.superuser_email,
                           password=self.superuser_password)
         response = self.client.post(self.merge_url)
@@ -3796,11 +3797,9 @@ class FacilityMergeTest(APITestCase):
         self.assertEqual(FacilityClaim.REVOKED, self.facility_2_claim.status)
         self.assertEqual(self.superuser,
                          self.facility_2_claim.status_change_by)
-        # Modifying tzinfo avoids
-        # TypeError: can't compare offset-naive and offset-aware datetimes
         self.assertGreater(
-            self.facility_2_claim.status_change_date.replace(tzinfo=None),
-            just_before_change.replace(tzinfo=None))
+            self.facility_2_claim.status_change_date,
+            just_before_change)
 
     def test_required_params(self):
         self.client.login(email=self.superuser_email,
@@ -4925,16 +4924,16 @@ class SearchByList(APITestCase):
         response = self.client.get(url)
         self.assertEqual(200, response.status_code)
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], self.facility.id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], self.facility.id)
 
     def test_fetched_by_list(self):
         url = '/api/facilities/?lists={}'.format(self.list.id)
         response = self.client.get(url)
         self.assertEqual(200, response.status_code)
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], self.facility.id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], self.facility.id)
 
     def test_fetched_by_list_and_contributor(self):
         url = '/api/facilities/?contributors={}&lists={}'.format(
@@ -4942,8 +4941,8 @@ class SearchByList(APITestCase):
         response = self.client.get(url)
         self.assertEqual(200, response.status_code)
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], self.facility.id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], self.facility.id)
 
 
 class UpdateLocationTest(FacilityAPITestCaseBase):
@@ -6507,7 +6506,7 @@ class FacilitySubmitTest(FacilityAPITestCaseBase):
         fields = NonstandardField.objects.filter(
             contributor=self.user.contributor).values_list(
             'column_name', flat=True)
-        self.assertEquals(1, len(fields))
+        self.assertEqual(1, len(fields))
         self.assertIn('extra_1', fields)
 
     def test_exact_matches_with_create_false(self):
@@ -7264,8 +7263,8 @@ class PPEFieldTest(TestCase):
                 'recall_weight': 0.5,
                 'code_version': 'abcd1234',
             },
-            'started': str(datetime.utcnow()),
-            'finished': str(datetime.utcnow()),
+            'started': str(timezone.now()),
+            'finished': str(timezone.now()),
         }
 
     def test_match_populates_ppe(self):
@@ -7656,7 +7655,7 @@ class CloseListTest(TestCase):
         self.assertFalse(f_two.is_closed)
 
         activity = FacilityActivityReport.objects.all().count()
-        self.assertEquals(2, activity)
+        self.assertEqual(2, activity)
 
 
 class NonstandardFieldsApiTest(APITestCase):
@@ -7686,7 +7685,7 @@ class NonstandardFieldsApiTest(APITestCase):
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
-        self.assertEquals(7, len(content))
+        self.assertEqual(7, len(content))
         self.assertIn('extra_1', content)
         self.assertIn('extra_2', content)
         self.assertIn('parent_company', content)
@@ -7697,7 +7696,7 @@ class NonstandardFieldsApiTest(APITestCase):
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
-        self.assertEquals(5, len(content))
+        self.assertEqual(5, len(content))
         self.assertNotIn('extra_1', content)
         self.assertNotIn('extra_2', content)
         self.assertIn('parent_company', content)
@@ -7819,13 +7818,13 @@ class ContributorFieldsApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
         contributor_fields = content['properties']['contributor_fields']
-        self.assertEquals(2, len(contributor_fields))
+        self.assertEqual(2, len(contributor_fields))
         field = contributor_fields[0]
         field_two = contributor_fields[1]
-        self.assertEquals('ExtraOne', field['label'])
-        self.assertEquals('data one', field['value'])
-        self.assertEquals('ExtraTwo', field_two['label'])
-        self.assertEquals(None, field_two['value'])
+        self.assertEqual('ExtraOne', field['label'])
+        self.assertEqual('data one', field['value'])
+        self.assertEqual('ExtraTwo', field_two['label'])
+        self.assertEqual(None, field_two['value'])
 
     def test_single_fields(self):
         response = self.client.get(
@@ -7834,13 +7833,13 @@ class ContributorFieldsApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
         contributor_fields = content['properties']['contributor_fields']
-        self.assertEquals(2, len(contributor_fields))
+        self.assertEqual(2, len(contributor_fields))
         field = contributor_fields[0]
         field_two = contributor_fields[1]
-        self.assertEquals('ExtraOne', field['label'])
-        self.assertEquals(None, field['value'])
-        self.assertEquals('ExtraTwo', field_two['label'])
-        self.assertEquals('data two', field_two['value'])
+        self.assertEqual('ExtraOne', field['label'])
+        self.assertEqual(None, field['value'])
+        self.assertEqual('ExtraTwo', field_two['label'])
+        self.assertEqual('data two', field_two['value'])
 
     def test_without_embed(self):
         response = self.client.get(
@@ -7849,7 +7848,7 @@ class ContributorFieldsApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
         contributor_fields = content['properties']['contributor_fields']
-        self.assertEquals(0, len(contributor_fields))
+        self.assertEqual(0, len(contributor_fields))
 
     def test_without_contributor(self):
         response = self.client.get(
@@ -7857,7 +7856,7 @@ class ContributorFieldsApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
         contributor_fields = content['properties']['contributor_fields']
-        self.assertEquals(0, len(contributor_fields))
+        self.assertEqual(0, len(contributor_fields))
 
     def test_inactive_match(self):
         self.match_one.is_active = False
@@ -7868,17 +7867,17 @@ class ContributorFieldsApiTest(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         content = json.loads(response.content)
         contributor_fields = content['properties']['contributor_fields']
-        self.assertEquals(2, len(contributor_fields))
+        self.assertEqual(2, len(contributor_fields))
         field = contributor_fields[0]
         field_two = contributor_fields[1]
-        self.assertEquals('ExtraOne', field['label'])
-        self.assertEquals(None, field['value'])
-        self.assertEquals('ExtraTwo', field_two['label'])
-        self.assertEquals(None, field_two['value'])
+        self.assertEqual('ExtraOne', field['label'])
+        self.assertEqual(None, field['value'])
+        self.assertEqual('ExtraTwo', field_two['label'])
+        self.assertEqual(None, field_two['value'])
 
     def test_custom_text(self):
         indexes = FacilityIndex.objects.count()
-        self.assertEquals(2, indexes)
+        self.assertEqual(2, indexes)
         index_one = FacilityIndex.objects.get(id=self.facility.id)
         index_one_data = '{}|data one'.format(self.contributor.id)
         self.assertIn(index_one_data, index_one.custom_text)
@@ -8123,7 +8122,7 @@ class ParentCompanyTestCase(FacilityAPITestCaseBase):
         }, ef.value)
         facility_index = FacilityIndex.objects.get(id=ef.facility.id)
         self.assertIn('A random value', facility_index.parent_company_name)
-        self.assertEquals(0, len(facility_index.parent_company_id))
+        self.assertEqual(0, len(facility_index.parent_company_id))
 
     @skip('Skip fuzzy matching. Will revisit at #1805')
     @patch('api.geocoding.requests.get')
@@ -8197,8 +8196,8 @@ class ParentCompanyTestCase(FacilityAPITestCaseBase):
             self.url + '?parent_company={}'.format(self.contributor.id)
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
     @patch('api.geocoding.requests.get')
     def test_search_by_id(self, mock_get):
@@ -8219,8 +8218,8 @@ class ParentCompanyTestCase(FacilityAPITestCaseBase):
             self.url + '?parent_company={}'.format(self.contributor.name)
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
     @patch('api.geocoding.requests.get')
     def test_search_by_multiple(self, mock_get):
@@ -8243,8 +8242,8 @@ class ParentCompanyTestCase(FacilityAPITestCaseBase):
             )
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
 
 class ProductTypeTestCase(FacilityAPITestCaseBase):
@@ -8358,8 +8357,8 @@ class ProductTypeTestCase(FacilityAPITestCaseBase):
 
         response = self.client.get(self.url + '?product_type=A')
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
 
 class FacilityAndProcessingTypeAPITest(FacilityAPITestCaseBase):
@@ -8485,8 +8484,8 @@ class FacilityAndProcessingTypeAPITest(FacilityAPITestCaseBase):
 
         response = self.client.get(self.url + '?processing_type=cutting')
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
     @patch('api.geocoding.requests.get')
     def test_search_by_facility_type(self, mock_get):
@@ -8507,8 +8506,8 @@ class FacilityAndProcessingTypeAPITest(FacilityAPITestCaseBase):
             self.url + '?facility_type=final%20product%20assembly'
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
 
 class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
@@ -8534,7 +8533,7 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
         self.assertEqual({'min': 2000, 'max': 2000}, ef.value)
 
         facility_index = FacilityIndex.objects.get(id=ef.facility.id)
-        self.assertEquals(1, len(facility_index.number_of_workers))
+        self.assertEqual(1, len(facility_index.number_of_workers))
         self.assertIn('1001-5000', facility_index.number_of_workers)
 
     @patch('api.geocoding.requests.get')
@@ -8555,7 +8554,7 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
         self.assertEqual({'min': 0, 'max': 500}, ef.value)
 
         facility_index = FacilityIndex.objects.get(id=ef.facility.id)
-        self.assertEquals(1, len(facility_index.number_of_workers))
+        self.assertEqual(1, len(facility_index.number_of_workers))
         self.assertIn('Less than 1000', facility_index.number_of_workers)
 
     @patch('api.geocoding.requests.get')
@@ -8576,7 +8575,7 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
         self.assertEqual({'min': 0, 'max': 10000}, ef.value)
 
         facility_index = FacilityIndex.objects.get(id=ef.facility.id)
-        self.assertEquals(3, len(facility_index.number_of_workers))
+        self.assertEqual(3, len(facility_index.number_of_workers))
         self.assertIn('Less than 1000', facility_index.number_of_workers)
         self.assertIn('1001-5000', facility_index.number_of_workers)
         self.assertIn('5001-10000', facility_index.number_of_workers)
@@ -8599,7 +8598,7 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
         self.assertEqual({'min': 20000, 'max': 100000}, ef.value)
 
         facility_index = FacilityIndex.objects.get(id=ef.facility.id)
-        self.assertEquals(1, len(facility_index.number_of_workers))
+        self.assertEqual(1, len(facility_index.number_of_workers))
         self.assertIn('More than 10000', facility_index.number_of_workers)
 
     @patch('api.geocoding.requests.get')
@@ -8622,8 +8621,8 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
             self.url + '?number_of_workers=1001-5000'
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
     @patch('api.geocoding.requests.get')
     def test_search_without_matches(self, mock_get):
@@ -8642,7 +8641,7 @@ class NumberOfWorkersAPITest(FacilityAPITestCaseBase):
             self.url + '?number_of_workers=More%20than%2010000'
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 0)
+        self.assertEqual(data['count'], 0)
 
 
 class NativeLanguageNameAPITest(FacilityAPITestCaseBase):
@@ -8671,8 +8670,8 @@ class NativeLanguageNameAPITest(FacilityAPITestCaseBase):
             self.url + '?native_language_name={}'.format(self.long_name)
         )
         data = json.loads(response.content)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
 
 class SectorAPITest(FacilityAPITestCaseBase):
@@ -8699,8 +8698,8 @@ class SectorAPITest(FacilityAPITestCaseBase):
         )
         data = json.loads(response.content)
         print(data)
-        self.assertEquals(data['count'], 1)
-        self.assertEquals(data['features'][0]['id'], facility_id)
+        self.assertEqual(data['count'], 1)
+        self.assertEqual(data['features'][0]['id'], facility_id)
 
 
 class ExactMatchTest(FacilityAPITestCaseBase):
@@ -8725,51 +8724,51 @@ class ExactMatchTest(FacilityAPITestCaseBase):
             'id': 1,
             'facility_id': 1,
             'source__contributor_id': self.contributor_2.id,
-            'updated_at': datetime.now(),
+            'updated_at': timezone.now(),
         }, {
             'id': 2,
             'facility_id': 2,
             'source__contributor_id': self.contributor.id,
-            'updated_at': datetime.now(),
+            'updated_at': timezone.now(),
         }]
         results = sort_exact_matches(matches, self.active_item_ids,
                                      self.contributor)
-        self.assertEquals(results[0]['facility_id'], 2)
-        self.assertEquals(results[1]['facility_id'], 1)
+        self.assertEqual(results[0]['facility_id'], 2)
+        self.assertEqual(results[1]['facility_id'], 1)
 
     def test_sorting_active(self):
         matches = [{
             'id': 1,
             'facility_id': 3,
             'source__contributor_id': self.contributor.id,
-            'updated_at': datetime.now(),
+            'updated_at': timezone.now(),
         }, {
             'id': 1,
             'facility_id': 1,
             'source__contributor_id': self.contributor.id,
-            'updated_at': datetime.now(),
+            'updated_at': timezone.now(),
         }]
         results = sort_exact_matches(matches, self.active_item_ids,
                                      self.contributor)
-        self.assertEquals(results[0]['facility_id'], 1)
-        self.assertEquals(results[1]['facility_id'], 3)
+        self.assertEqual(results[0]['facility_id'], 1)
+        self.assertEqual(results[1]['facility_id'], 3)
 
     def test_sorting_newest(self):
         matches = [{
             'id': 1,
             'facility_id': 2,
             'source__contributor_id': self.contributor.id,
-            'updated_at': datetime.now().replace(year=2020),
+            'updated_at': timezone.now().replace(year=2020),
         }, {
             'id': 1,
             'facility_id': 1,
             'source__contributor_id': self.contributor.id,
-            'updated_at': datetime.now(),
+            'updated_at': timezone.now(),
         }]
         results = sort_exact_matches(matches, self.active_item_ids,
                                      self.contributor)
-        self.assertEquals(results[0]['facility_id'], 1)
-        self.assertEquals(results[1]['facility_id'], 2)
+        self.assertEqual(results[0]['facility_id'], 1)
+        self.assertEqual(results[1]['facility_id'], 2)
 
 
 class FacilityDetailSerializerTest(TestCase):